requirements.txt
ffmpeg-python
pydub
openai-whisper
deep_translator #googletrans==4.0.0-rc1
gTTS
soundfile
librosa
psola
numpy
scipy
transformers
torch
